Functions files
Takes snippets of code and runs it in a function. This is to run before all other code that would use functions from this file. After running, the long lines of code can be simplified down into a singular line and sections, such as the packets and libraries, can be called eaily whenever needed.